+++ src/greenlet/greenlet_compiler_compat.hpp
@@ -46,7 +46,6 @@
 #  define G_FP_TMPL_STATIC
 # endif
 
-
 #    define G_NO_COPIES_OF_CLS(Cls) private:     \
     Cls(const Cls& other) = delete; \
     Cls& operator=(const Cls& other) = delete