+++ CHANGES.rst
@@ -26,7 +26,9 @@
 - Remove support for end-of-life Python versions, including Python
   2.7, Python 3.5 and Python 3.6.
 - Require a compiler that supports ``noinline`` directives. See
-  `issue 271 <https://github.com/python-greenlet/greenlet/issues/266>`_
+  `issue 271
+  <https://github.com/python-greenlet/greenlet/issues/266>`_.
+- Require a compiler that supports C++11.
 
 
 2.0.2 (2023-01-28)